feat(Car 메소드 테스트) : 입력 유효성 검사 및 go() 테스트
 - 자동차 이름이 빈 칸일때의 유효성 검사 추가
 - go() 테스트 추가
 - CarRacingProgram에 있던 자동차 이름 유효성 검사 메소드 Car 도메인으로 이동
