+++ src/test/java/carracing/CarTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class CarTest {
 
@@ -23,5 +24,24 @@ public class CarTest {
     @Test
     void positionTest(){
         assertThat(car.getPosition()).isEqualTo(2);
+    }
+
+    @Test
+    void carNameLengthTest(){
+        assertThatThrownBy(() ->{
+            new Car("car2car3");
+        }).isInstanceOf(RuntimeException.class);
+    }
+
+    @Test
+    void carNameBlankTest(){
+        assertThatThrownBy(() ->{
+            new Car("");
+        }).isInstanceOf(RuntimeException.class);
+    }
+
+    @Test
+    void goTest(){
+        assertThat(new Car("car1", 2).go()).isEqualTo(3);
     }
 }
